Make base wrapper untrainable
There are situations where Tensor/Input is used as parameter variables,
it is possible that they are fed to gradient function. So as not to crash,
need to add `triainable` attribtue.

Out of all wrapper types (Input/Variable/Tensor), only explicitly marked
Variables are trainable, so default to trainable=False.